Fix Makefile, snapcraft.yaml file, and icon image
Change the Makefile default to use Ubuntu OpenJDK 11 from the Main
repository instead of the old OpenJDK 14 from the Universe repository.

Get the source from the local project containing the 'snapcraft.yaml'
file instead of from the GitHub repository.

Delete empty directories only if they're also found in the base snap.

Use the variable '$SNAPCRAFT_ARCH_TRIPLET' in the library paths instead
of 'x86_64-linux-gnu', which is specific to the amd64 architecture.

Add a white outline to the Duke icon image so that it looks good when
displayed on either a light or a dark background.