Add support for building on Fedora-based systems
Add support to the Makefile for building on Fedora. For systems
based on either Debian or Fedora, add comments to the Makefile
documenting the build requirements, and add sample files exporting
the system-specific environment variables JAVA_HOME, JUNIT4, HAMCREST,
JAVAFX_LIB, and JAVAFX_MOD.

Use the more portable machine hardware name (uname -m) instead of
the Debian architecture in the file name of the compressed archive.